# src/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "",
        "version": "1.0.0",
        "title": "Jukebox",
        "contact": {
            "name": "API Support",
            "email": "apiteam@swagger.io"
        }
    },
    "tags": [
        {
            "name": "genius",
            "description": "accessing Genius API endpoints",
            "externalDocs": {
                "description": "Find out more",
                "url": "https://docs.genius.com/"
            }
        },
        {
            "name": "Spotify Web API",
            "description": "accessing Spotify Web API endpoints",
            "externalDocs": {
                "description": "Find out more",
                "url": "https://developer.spotify.com/documentation/web-api/reference/#/"
            }
        }
    ],
    "paths": {
        "/lyrics/{title}/{artist}": {
            "get": {
                "tags": [
                    "genius"
                ],
                "description": "get song lyrics by searching for title and artist",
                "parameters": [
                    {
                        "in": null,
                        "name": "title",
                        "description": "title of the song",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "artist",
                        "description": "main artist of the song",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song lyrics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "items": {
                                        "$ref": "#/components/schemas/lyric"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lyrics not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "items": {
                                        "$ref": "#/components/schemas/error"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/playlists/{playlist_id}": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "get content of playlist by id e.g. 7LBhiiutTkDBGUfBzVPaoU",
                "parameters": [
                    {
                        "in": "path",
                        "name": "playlist_id",
                        "description": "playlist id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A playlist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Track"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    },
                    "404": {
                        "description": "Service not found"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "requests user authorization to access data @try it out doesn't work in Swagger-UI",
                "responses": {
                    "200": {
                        "description": "redirects to Spotify Accounts Service, login prompt"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    }
                }
            }
        },
        "/callback": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "requests access and refresh token, is called after successful login @try it out doesn't work in Swagger-UI",
                "responses": {
                    "200": {
                        "description": "sets access and refresh token"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    }
                }
            }
        },
        "/player/pause": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "pause current playback",
                "responses": {
                    "200": {
                        "description": "successfully paused"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    },
                    "403": {
                        "description": "Forbidden, player is already paused"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            }
        },
        "/player/play": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "resume current playback",
                "responses": {
                    "200": {
                        "description": "successfully played"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    },
                    "403": {
                        "description": "Forbidden, player is already playing"
                    },
                    "404": {
                        "description": "Device not found"
                    }
                }
            }
        },
        "/player/queue/{track_id}": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "add track to queue by id e.g. 64BbK9SFKH2jk86U3dGj2P",
                "parameters": [
                    {
                        "in": "path",
                        "name": "track_id",
                        "description": "track id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully added to queue"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    }
                }
            }
        },
        "/player": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "get information about the user's current player and playback state",
                "responses": {
                    "200": {
                        "description": "current playback state and current song",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/currentTrack"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    },
                    "404": {
                        "description": "no active device found"
                    }
                }
            }
        },
        "/player/devices": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "user's available devices",
                "responses": {
                    "200": {
                        "description": "available devices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Device"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    }
                }
            }
        },
        "/switchPlayer/{device_id}": {
            "get": {
                "tags": [
                    "Spotify Web API"
                ],
                "description": "transfer playback to the given device id and start playback",
                "parameters": [
                    {
                        "in": "path",
                        "name": "device_id",
                        "description": "device id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "playback switched"
                    },
                    "401": {
                        "description": "Bad or expired token. This can happen if the user revoked a token or the access token has expired. You should re-authenticate the user."
                    },
                    "404": {
                        "description": "device not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "lyric": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "lyrics": {
                        "type": "string"
                    }
                }
            },
            "error": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "Track": {
                "type": "object",
                "properties": {
                    "track": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "artists": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "album": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    }
                }
            },
            "Playlist": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "tracks": {
                        "type": "integer"
                    },
                    "owner": {
                        "type": "string"
                    }
                }
            },
            "currentTrack": {
                "type": "object",
                "properties": {
                    "track": {
                        "type": "string"
                    },
                    "track_id": {
                        "type": "string"
                    },
                    "device": {
                        "type": "string"
                    },
                    "device_id": {
                        "type": "string"
                    },
                    "artists": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "album": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    },
                    "progress": {
                        "type": "string"
                    },
                    "isPlayed": {
                        "type": "boolean"
                    }
                }
            },
            "Device": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "is_private_session": {
                        "type": "boolean"
                    },
                    "is_restricted": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "volume_percent": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}